[MK/MIN-236] add API call
